requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
numpy==1.26.4
psutil==6.0.0
pyarrow==17.0.0
PyGetWindow==0.0.9
PyJWT==2.9.0
pylance==0.16.1
pymongo==4.8.0
PyRect==0.2.0
python-xlib==0.33
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
xlib==0.21
